requirements.txt
python-telegram-bot==20.3
qdrant-client==1.6.4
openai==1.3.5
notion-client==2.2.0
google-api-python-client==2.108.0
python-dotenv==1.0.0
